pyproject.toml
[tool.poetry]
name = "aicustomify-backend"
version = "0.0.0"
description = ""
authors = ["Edgar Valdez <edgarvaldez99@gmail.com>"]
packages = [
    { include = "app" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
python-decouple = "^3.8"
pymongo = "^4.8.0"
openai = "^1.41.0"
pandas = "^2.2.2"
numpy = "1.26.4"
langchain = "^0.2.14"
faiss-cpu = "^1.8.0.post1"
langchain-community = "^0.2.12"
sentence-transformers = "^3.0.1"
langchain-huggingface = "^0.0.3"
diskcache = "^5.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
